updated scripts

+++ install/install.sh
@@ -4,7 +4,9 @@ export SUDO_ASKPASS=${HOME}/Unitree/autostart/passwd.sh
 autostart_dir=${HOME}/.config/autostart
 cp files/ice9_autostart.sh ${autostart_dir}/ice9_autostart.sh.desktop
 
-mv ${autostart_dir}/update.sh.desktop ${autostart_dir}/update.sh.desktop.bak
+if [ -f ${autostart_dir}/update.sh.desktop ]; then
+    mv ${autostart_dir}/update.sh.desktop ${autostart_dir}/update.sh.desktop.bak
+fi
 if [ -f ${autostart_dir}/slam_planner.sh.desktop ]; then
     mv ${autostart_dir}/slam_planner.sh.desktop ${autostart_dir}/slam_planner.sh.desktop.bak
 fi
